feat: add inline task editing functionality to todo list
- Add inline edit form display within TodoList component
- Implement edit state management with visual highlight ring
- Fix dynamic API URL resolution for SSR/client-side differences
- Update tests to match new DOM structure with padding classes
- All tests passing (390/390)
